% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{aggregate_trackedCells}
\alias{aggregate_trackedCells}
\title{Aggregate trackedCells Objects}
\usage{
aggregate_trackedCells(x, ..., meta_id_field = c("tiff_file",
  "experiment", "condition", "replicate"))
}
\arguments{
\item{x}{a trackedCells-class object where cells have already been tracked}

\item{...}{one or more trackedCells-class object(s) where cells have already been tracked}

\item{meta_id_field}{string, can take one of the following values, c("tiff_file", "experiment", 
"condition", "replicate"). Indicates the meta-data column used as unique ID for the image/experiment.
Can be abbreviated. Defaults to "tiff_file".}
}
\value{
An aggregate data.frame including all cells that were tracked over two or more images/experiments. 
The data.frame includes the following columns: "new.ID", "frame.ID", "X", "Y", "cell.ID", "tiff_name", 
"experiment", "condition", "replicate". The "new.ID" uniquely identifies a cell in a given image/experiment.
}
\description{
Aggregate two or more trackedCells-class objects together. Input objects must carry information of cell
tracks (oterwise an error will be raised). All tracks form the different experiments/images are returned in a
large data.frame. A new unique ID is assigned to specifically identify each cell track from each image/experiment.
}
\details{
each trackedCells-class object passed to this function requires a unique identifier (such as a unique 
tiff_file name). Any of the metadata columns can be used as unique ID for an image/experiment. The function
will raise an error if non-unique identifiers are found across the input objects.
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
