% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{bpass}
\alias{bpass}
\title{Perform a bandpass by convolving with an appropriate kernel}
\usage{
bpass(image_array, lnoise, lobject = NULL, threshold)
}
\arguments{
\item{image_array}{Numeric matrix corresponding to the image to be filtered}

\item{lnoise}{Characteristic lengthscale of noise in pixels.}

\item{lobject}{Integer length in pixels somewhat larger than a typical object}

\item{threshold}{By default, after the convolution, any negative pixels are reset to 0.
Threshold changes the threshhold for setting pixels to 0. Positive values may be useful for removing
stray noise or small particles.}
}
\value{
Numeric matrix corresponding to the filtered image
}
\description{
Implements a real-space bandpass filter that suppresses pixel noise and long-wavelength 
image variations while retaining information of a characteristic size.
First, a lowpassed image is produced by convolving the original with a gaussian.  
Next, a second lowpassed image is produced by convolving the original with a 
boxcar function. By subtracting the boxcar version from the gaussian version, 
we are using the boxcar version to perform a highpass.
This code 'bpass.pro' is copyright 1997, John C. Crocker and 
David G. Grier.  It should be considered 'freeware'- and may be
distributed freely in its original form when properly attributed.
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
