% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\docType{package}
\name{cellTracker-package}
\alias{cellTracker}
\alias{cellTracker-package}
\title{Track Fluorescent Cells and Compute Migration Stats}
\description{
Track Fluorescent Cells and Compute Migration Stats
}
\keyword{internal}
