% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{cntrd}
\alias{cntrd}
\title{Calculates Centroids}
\usage{
cntrd(im, mx, sz, interactive = NULL)
}
\arguments{
\item{im}{numeric matrix corresponding to the image to process}

\item{mx}{location of local maxima to pixel-levels accuracy}

\item{sz}{diamter of the window over which to average to calculate the centroid. should be big enough.}

\item{interactive}{numeric; if set to 1 (or any positive number), an image showing the 
computed centroids will be visualized}
}
\value{
a data.frame with 4 columns, containing, x, y, brightness, and 
the square of the radius of gyration for each cell.
}
\description{
Calculates the centroid of bright spots to sub-pixel accuracy. Inspired by Grier & 
Crocker's feature for IDL, but greatly simplified and optimized for matlab, and then 
further ported to R. CREATED: Eric R. Dufresne, Yale University, Feb 4 2005.
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
