% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{compute_tracks_stats}
\alias{compute_tracks_stats}
\title{Compute Tracks Stats}
\usage{
compute_tracks_stats(tc_obj, time_between_frames,
  resolution_pixel_per_micron)
}
\arguments{
\item{tc_obj}{a trackedCells object}

\item{time_between_frames}{integer, time interval between two successive frames were taken}

\item{resolution_pixel_per_micron}{integer, image resolution, i.e. number of pixels per micron}
}
\value{
a trackedCells object, including cell track statistics
}
\description{
Wrapper for the migrationStats() function. It computes statistics for a 
trackedCells object where cells have already been tracked.
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
