% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{detect_radii}
\alias{detect_radii}
\title{Detect Linear Paricle Diameters}
\usage{
detect_radii(x)
}
\arguments{
\item{x}{numeric or logical vector}
}
\value{
data.frame including two columns: MPOS indicates the centroid position of a particle, 
and LEN indicates the diameter size
}
\description{
Estimates the diameters of particles in a numeric or logical vector
}
\examples{
cellTracker:::detect_radii(c(0, 0, 1, 1, 0, 1, 1, 1, 1, 0, 0, 1, 0, 0, 1, 1))

}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
