% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{get_image_stacks}
\alias{get_image_stacks}
\title{Get Image Stacks}
\usage{
get_image_stacks(tc_obj)
}
\arguments{
\item{tc_obj}{a trackedCells object}
}
\value{
a list including stack images (formatted as numeric matrices)
}
\description{
Extract Images Stacks from a trackedCells object
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
