% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{get_population_stats}
\alias{get_population_stats}
\title{Get Cell population stats}
\usage{
get_population_stats(tc_obj)
}
\arguments{
\item{tc_obj}{a trackedCells object}
}
\value{
data.frame including cell population stats
}
\description{
Extract cell population statistics from a trackedCells object
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
