% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{get_tracks}
\alias{get_tracks}
\title{Get Track Data}
\usage{
get_tracks(tc_obj, attach_meta = FALSE)
}
\arguments{
\item{tc_obj}{a trackedCells object}

\item{attach_meta}{logical, shall metaData be attached to tracks}
}
\value{
a data.frame including cell tracks data
}
\description{
Extract Track Data from a trackedCells object
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
