% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{load_tif}
\alias{load_tif}
\title{Import Image from TIFF}
\usage{
load_tif(tiff_file, experiment = NULL, condition = NULL,
  replicate = NULL)
}
\arguments{
\item{tiff_file}{path to a TIFF file to be read in}

\item{experiment}{string, a label to describe the experiment (optional). Can be NULL}

\item{condition}{string, a label to describe the experimental condition (optional). Can be NULL}

\item{replicate}{string, a label to identify the replicate (optional). Can be NULL}
}
\value{
a trackedCells object
}
\description{
Import a .tif stack containing fluorescently labeled point particles to be tracked
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
