% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{matfix}
\alias{matfix}
\title{Clean And Reformat a Numeric Matrix}
\usage{
matfix(x)
}
\arguments{
\item{x}{matrix or data.frame including numeric data (or data that can be coerced to integer)}
}
\value{
numeric matrix with all its elements coerced to integer
}
\description{
Convert any matrix-lie object to a numeric Matrix, and coerces all the elements to integer. 
Row names and column names are removed
}
\examples{
tmp <- data.frame(A = 1:4, B=c(3.1, 2.8, 3.3, 9.1), C = FALSE)
cellTracker:::matfix(tmp)

}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
