% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{migrationStats}
\alias{migrationStats}
\title{Compute Cell Migration Statistics}
\usage{
migrationStats(tracks, interval_time, pixel_micron)
}
\arguments{
\item{tracks}{data.frame with cell tracks information}

\item{interval_time}{integer, time interval between two successive frames were taken}

\item{pixel_micron}{integer, image resolution, i.e. number of pixels per micron}
}
\value{
list of stats calculated for the cell tracks. Info include variables of speed, 
distance, euclidean displacement, persistence, angular displacement, 
yFMI, xFMI, y-displacement, x-displacement and frames
}
\description{
Calculate the statistics from X/Y positional data obtained from cell tracks
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
