% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{next_odd}
\alias{next_odd}
\title{Return the Next Odd Integer}
\usage{
next_odd(x)
}
\arguments{
\item{x}{a vector of class numeric}
}
\value{
a vector of class integer
}
\description{
Returns the smallest odd number bigger than the number(s) provided as the argument
}
\examples{
cellTracker:::next_odd(2:5)

}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
