% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\docType{data}
\name{sample_tracked_cells}
\alias{sample_tracked_cells}
\title{Sample Stack of Fluorescent Cells}
\format{An object of class \code{trackedCells} of length 1.}
\usage{
sample_tracked_cells
}
\description{
Sample Stack of Fluorescent Cells to be used for computing cell tracks and stats
}
\keyword{internal}
