% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{set_cells_meta}
\alias{set_cells_meta}
\title{Set MetaData}
\usage{
set_cells_meta(tc_obj, experiment = NULL, condition = NULL,
  replicate = NULL)
}
\arguments{
\item{tc_obj}{a trackedCells object}

\item{experiment}{string, a label to describe the experiment (optional). Can be NULL}

\item{condition}{string, a label to describe the experimental condition (optional). Can be NULL}

\item{replicate}{string, a label to identify the replicate (optional). Can be NULL}
}
\value{
a list including three items: experiment name, condition label, and replicate ID.
}
\description{
Write/Replace MetaData of a trackedCells object
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
