% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{track}
\alias{track}
\title{Track cells}
\usage{
track(xyzs, maxdisp, params)
}
\arguments{
\item{xyzs}{an array listing the scrambled coordinates and data of the 
different particles at different times}

\item{maxdisp}{an estimate of the maximum distance that a particle would 
move in a single time interval}

\item{params}{a list containing a few tracking parameters that are needed 
for the analysis}
}
\value{
data.frame including cell tracks data
}
\description{
Constructs n-dimensional trajectories from a scrambled list of particle 
coordinates determined at discrete times (e.g. in consecutive image frames)
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{No}
\keyword{Warning-}
\keyword{for}
\keyword{found}
\keyword{internal}
\keyword{positions}
\keyword{t='}
