% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{visualize_cntr}
\alias{visualize_cntr}
\title{Visualize Centroids}
\usage{
visualize_cntr(centroids, width_px, height_px, pnt.cex = 1.2,
  txt.cex = 0.9, offset = 0.18, col = "red2")
}
\arguments{
\item{centroids}{centroid data.frame}

\item{width_px}{width of the image in pixels}

\item{height_px}{height of the image in pixels}

\item{pnt.cex}{cex of the point (circle) drawn around each cell}

\item{txt.cex}{cex of the text used to annotate the image}

\item{offset}{offset for the text annotations}

\item{col}{color of the points, e.g. "red2"}
}
\value{
None
}
\description{
Annotates centroids over an image
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
