% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{visualize_img}
\alias{visualize_img}
\title{Visualize a matrix image}
\usage{
visualize_img(img_mtx, col = NULL, ...)
}
\arguments{
\item{img_mtx}{numeric matrix corresponding to a image}

\item{col}{character vector corresponding to a valid color palette}

\item{...}{additional arguments will be passed to graphics::image()}
}
\value{
None
}
\description{
Shows an image representation of a numeric matrix. Typically, this is a non-negative numeric matrix, 
where signal (high values) corresponds to the presence of cells, or cell-like particles
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{cellTracker}
\keyword{internal}
