% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTracker_core.R
\name{visualize_stack_centroids}
\alias{visualize_stack_centroids}
\title{Visualize Cells in an Image Stack}
\usage{
visualize_stack_centroids(tc_obj, stack = 1, pnt.cex = 1.2,
  txt.cex = 0.9, offset = 0.18, main = NULL)
}
\arguments{
\item{tc_obj}{a trackedCells object}

\item{stack}{index of the image stack to use}

\item{pnt.cex}{cex of the points drawn around cells}

\item{txt.cex}{cex of the text used to annotate cells}

\item{offset}{offset value for the annotation}

\item{main}{string used for the plot title, can be NULL= NULL}
}
\value{
None
}
\description{
Visualize objects that were identified as cells in a given image stack
}
\references{
\url{https://www.data-pulse.com/dev_site/celltracker/}
\url{https://www.mathworks.com/matlabcentral/fileexchange/60349-fasttracks}
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
